Enable classes to be in packages
