--- src/main/java/AjException.java
@@ -1,5 +1,0 @@
-public class AjException extends Exception{
-  AjException(String msg) {
-    super(msg);
-  }
-}
