--- src/main/java/IndexOutOfRangeException.java
@@ -1,5 +1,0 @@
-public class IndexOutOfRangeException extends AjException {
-  IndexOutOfRangeException(int val) {
-    super("OOPS!!! Please choose an index within the number of items in the list " + "(" + val + ")");
-  }
-}
