+++ src/main/java/aj/AjException.java
@@ -1,0 +1,7 @@
+package aj;
+
+public class AjException extends Exception {
+    AjException(String msg) {
+        super(msg);
+    }
+}
