+++ src/main/java/aj/EmptyDescriptionException.java
@@ -1,0 +1,9 @@
+package aj;
+
+public class EmptyDescriptionException extends AjException {
+    EmptyDescriptionException(String cmd, String helpMsg) {
+        super("OOPS!!! The description of a " + cmd + " cannot be empty.\n" +
+                "Try:\n" +
+                helpMsg);
+    }
+}
