+++ src/main/java/aj/IndexOutOfRangeException.java
@@ -1,0 +1,7 @@
+package aj;
+
+public class IndexOutOfRangeException extends AjException {
+    IndexOutOfRangeException(int val) {
+        super("OOPS!!! Please choose an index within the number of items in the list " + "(" + val + ")");
+    }
+}
