+++ src/main/java/aj/Ui.java
@@ -1,3 +1,5 @@
+package aj;
+
 public class Ui {
     TaskList taskList;
